+++ unit-tests/state/transit-gateways.test.js
@@ -248,6 +248,7 @@ describe("transit_gateways", () => {
         resource_group: "management-rg",
         global: false,
         connections: [{ tgw: "tg-test", vpc: "management" }],
+        gre_tunnels: [],
       };
       assert.deepEqual(
         state.store.json.transit_gateways[1],
@@ -279,6 +280,7 @@ describe("transit_gateways", () => {
               vpc: "workload",
             },
           ],
+          gre_tunnels: [],
         },
       ];
       assert.deepEqual(
@@ -621,6 +623,441 @@ describe("transit_gateways", () => {
               crns: ["crn:v1:bluemix:public:abcdf", "mooseeeeeeeeeeeeeeeeee"],
             })
           );
+        });
+      });
+    });
+  });
+  describe("transit_gateways.gre_tunnels", () => {
+    describe("transit_gateways.gre_tunnels.create", () => {
+      it("should create a gre tunnel", () => {
+        let state = newState();
+        state.update();
+        state.transit_gateways.gre_tunnels.create(
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          {
+            innerFormProps: {
+              arrayParentName: "transit-gateway",
+            },
+          }
+        );
+        assert.deepEqual(
+          state.store.json.transit_gateways[0].gre_tunnels[0],
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: null,
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          "it should remove unfound gateway and create tunnel"
+        );
+      });
+      it("should create a gre tunnel with gateway", () => {
+        let state = newState();
+        state.classic_gateways.create({
+          name: "gw",
+          hostname: "gw-host",
+          datacenter: "dal10",
+          network_speed: "1000",
+          private_network_only: false,
+          tcp_monitoring: false,
+          redundant_network: true,
+          public_bandwidth: 5000,
+          memory: 64,
+          notes: "Notes",
+          ipv6_enabled: false,
+          package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+          os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+          process_key_name: "INTEL_XEON_4210_2_20",
+          private_vlan: "vsrx-private",
+          public_vlan: "vsrx-public",
+          ssh_key: "example-classic",
+          disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+          hadr: false,
+        });
+        state.transit_gateways.gre_tunnels.create(
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          {
+            innerFormProps: {
+              arrayParentName: "transit-gateway",
+            },
+          }
+        );
+        assert.deepEqual(
+          state.store.json.transit_gateways[0].gre_tunnels[0],
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          "it should create tunnel"
+        );
+      });
+    });
+    describe("transit_gateways.gre_tunnels.save", () => {
+      it("should update a gre tunnel with gateway", () => {
+        let state = newState();
+        state.classic_gateways.create({
+          name: "gw",
+          hostname: "gw-host",
+          datacenter: "dal10",
+          network_speed: "1000",
+          private_network_only: false,
+          tcp_monitoring: false,
+          redundant_network: true,
+          public_bandwidth: 5000,
+          memory: 64,
+          notes: "Notes",
+          ipv6_enabled: false,
+          package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+          os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+          process_key_name: "INTEL_XEON_4210_2_20",
+          private_vlan: "vsrx-private",
+          public_vlan: "vsrx-public",
+          ssh_key: "example-classic",
+          disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+          hadr: false,
+        });
+        state.transit_gateways.gre_tunnels.create(
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          {
+            innerFormProps: {
+              arrayParentName: "transit-gateway",
+            },
+          }
+        );
+        state.transit_gateways.gre_tunnels.save(
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 2,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          {
+            arrayParentName: "transit-gateway",
+            data: {
+              gateway: "gw",
+            },
+          }
+        );
+        assert.deepEqual(
+          state.store.json.transit_gateways[0].gre_tunnels[0],
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 2,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          "it should create tunnel"
+        );
+      });
+    });
+    describe("transit_gateways.gre_tunnels.delete", () => {
+      it("should delete a gre tunnel with gateway", () => {
+        let state = newState();
+        state.classic_gateways.create({
+          name: "gw",
+          hostname: "gw-host",
+          datacenter: "dal10",
+          network_speed: "1000",
+          private_network_only: false,
+          tcp_monitoring: false,
+          redundant_network: true,
+          public_bandwidth: 5000,
+          memory: 64,
+          notes: "Notes",
+          ipv6_enabled: false,
+          package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+          os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+          process_key_name: "INTEL_XEON_4210_2_20",
+          private_vlan: "vsrx-private",
+          public_vlan: "vsrx-public",
+          ssh_key: "example-classic",
+          disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+          hadr: false,
+        });
+        state.transit_gateways.gre_tunnels.create(
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          {
+            innerFormProps: {
+              arrayParentName: "transit-gateway",
+            },
+          }
+        );
+        state.transit_gateways.gre_tunnels.delete(
+          {},
+          {
+            arrayParentName: "transit-gateway",
+            data: {
+              gateway: "gw",
+            },
+          }
+        );
+        assert.deepEqual(
+          state.store.json.transit_gateways[0].gre_tunnels,
+          [],
+          "it should create tunnel"
+        );
+      });
+      it("should delete a gre tunnel with null gateway", () => {
+        let state = newState();
+        state.classic_gateways.create({
+          name: "gw",
+          hostname: "gw-host",
+          datacenter: "dal10",
+          network_speed: "1000",
+          private_network_only: false,
+          tcp_monitoring: false,
+          redundant_network: true,
+          public_bandwidth: 5000,
+          memory: 64,
+          notes: "Notes",
+          ipv6_enabled: false,
+          package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+          os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+          process_key_name: "INTEL_XEON_4210_2_20",
+          private_vlan: "vsrx-private",
+          public_vlan: "vsrx-public",
+          ssh_key: "example-classic",
+          disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+          hadr: false,
+        });
+        state.classic_gateways.create({
+          name: "gw2",
+          hostname: "gw-host",
+          datacenter: "dal10",
+          network_speed: "1000",
+          private_network_only: false,
+          tcp_monitoring: false,
+          redundant_network: true,
+          public_bandwidth: 5000,
+          memory: 64,
+          notes: "Notes",
+          ipv6_enabled: false,
+          package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+          os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+          process_key_name: "INTEL_XEON_4210_2_20",
+          private_vlan: "vsrx-private",
+          public_vlan: "vsrx-public",
+          ssh_key: "example-classic",
+          disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+          hadr: false,
+        });
+        state.transit_gateways.gre_tunnels.create(
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: "gw",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          {
+            innerFormProps: {
+              arrayParentName: "transit-gateway",
+            },
+          }
+        );
+        state.transit_gateways.gre_tunnels.create(
+          {
+            tgw: "transit-gateway",
+            remote_bgp_asn: 12345,
+            zone: 1,
+            gateway: "gw2",
+            local_tunnel_ip: "1.2.3.4",
+            remote_tunnel_ip: "1.2.3.4",
+          },
+          {
+            innerFormProps: {
+              arrayParentName: "transit-gateway",
+            },
+          }
+        );
+        state.classic_gateways.delete({}, { data: { name: "gw" } });
+        state.transit_gateways.gre_tunnels.delete(
+          {},
+          {
+            arrayParentName: "transit-gateway",
+            data: {
+              gateway: null,
+            },
+          }
+        );
+        assert.deepEqual(
+          state.store.json.transit_gateways[0].gre_tunnels,
+          [
+            {
+              gateway: "gw2",
+              local_tunnel_ip: "1.2.3.4",
+              remote_bgp_asn: 12345,
+              remote_tunnel_ip: "1.2.3.4",
+              tgw: "transit-gateway",
+              zone: 1,
+            },
+          ],
+          "it should create tunnel"
+        );
+      });
+    });
+    describe("transit_gateways.gre_tunnels.schema", () => {
+      describe("transit_gateways.gre_tunnels.schema.gateway", () => {
+        describe("transit_gateways.gre_tunnels.schema.gateway.groups", () => {
+          it("should return groups", () => {
+            let craig = newState();
+            craig.classic_gateways.create({
+              name: "gw",
+              hostname: "gw-host",
+              datacenter: "dal10",
+              network_speed: "1000",
+              private_network_only: false,
+              tcp_monitoring: false,
+              redundant_network: true,
+              public_bandwidth: 5000,
+              memory: 64,
+              notes: "Notes",
+              ipv6_enabled: false,
+              package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+              os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+              process_key_name: "INTEL_XEON_4210_2_20",
+              private_vlan: "vsrx-private",
+              public_vlan: "vsrx-public",
+              ssh_key: "example-classic",
+              disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+              hadr: false,
+            });
+            craig.transit_gateways.gre_tunnels.create(
+              {
+                tgw: "transit-gateway",
+                remote_bgp_asn: 12345,
+                zone: 1,
+                gateway: "gw",
+                local_tunnel_ip: "1.2.3.4",
+                remote_tunnel_ip: "1.2.3.4",
+              },
+              {
+                innerFormProps: {
+                  arrayParentName: "transit-gateway",
+                },
+              }
+            );
+            assert.deepEqual(
+              craig.transit_gateways.gre_tunnels.gateway.groups(
+                {},
+                {
+                  craig: craig,
+                  arrayParentName: "transit-gateway",
+                  data: { gateway: "gw" },
+                }
+              ),
+              ["gw"],
+              "it should return list of gateways"
+            );
+          });
+          it("should return groups and filter out gateways with a GRE tunnel to the current gateway", () => {
+            let craig = newState();
+            craig.classic_gateways.create({
+              name: "gw",
+              hostname: "gw-host",
+              datacenter: "dal10",
+              network_speed: "1000",
+              private_network_only: false,
+              tcp_monitoring: false,
+              redundant_network: true,
+              public_bandwidth: 5000,
+              memory: 64,
+              notes: "Notes",
+              ipv6_enabled: false,
+              package_key_name: "VIRTUAL_ROUTER_APPLIANCE_1_GPBS",
+              os_key_name: "OS_JUNIPER_VSRX_19_4_UP_TO_1GBPS_STANDARD_SRIOV",
+              process_key_name: "INTEL_XEON_4210_2_20",
+              private_vlan: "vsrx-private",
+              public_vlan: "vsrx-public",
+              ssh_key: "example-classic",
+              disk_key_names: ["HARD_DRIVE_2_00_TB_SATA_2"],
+              hadr: false,
+            });
+            craig.transit_gateways.create({
+              name: "global",
+              global: true,
+              connections: [],
+            });
+            assert.deepEqual(
+              craig.transit_gateways.gre_tunnels.gateway.groups(
+                {},
+                { craig: craig, arrayParentName: "transit-gateway" }
+              ),
+              ["gw"],
+              "it should return list of gateways with already attached gateways filtered out"
+            );
+          });
+        });
+      });
+      describe("transit_gateways.gre_tunnels.schema.local_tunnel_ip", () => {
+        describe("transit_gateways.gre_tunnels.schema.local_tunnel_ip.invalid", () => {
+          it("should return true if not an ipv4 address", () => {
+            let craig = newState();
+            assert.isTrue(
+              craig.transit_gateways.gre_tunnels.local_tunnel_ip.invalid({
+                local_tunnel_ip: "aa",
+              }),
+              "it should be invalid"
+            );
+          });
+          it("should return true if an ipv4 cidr address", () => {
+            let craig = newState();
+            assert.isTrue(
+              craig.transit_gateways.gre_tunnels.local_tunnel_ip.invalid({
+                local_tunnel_ip: "10.10.10.10/10",
+              }),
+              "it should be invalid"
+            );
+          });
+        });
+        describe("transit_gateways.gre_tunnels.schema.local_tunnel_ip.invalidText", () => {
+          it("should return invalid text", () => {
+            let craig = newState();
+            assert.deepEqual(
+              craig.transit_gateways.gre_tunnels.local_tunnel_ip.invalidText(),
+              "Enter a valid IP address",
+              "it should be invalid"
+            );
+          });
         });
       });
     });